best-server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
oracledb
pymysql
jinja2
sqlalchemy
